+++ api/onnx_web/diffusion/run.py
@@ -12,7 +12,8 @@ from onnx_web.chain.base import ChainProgress
 from ..chain import upscale_outpaint
 from ..output import save_image, save_params
 from ..params import Border, ImageParams, Size, StageParams, UpscaleParams
-from ..server import JobContext, ServerContext
+from ..worker import WorkerContext
+from ..server import ServerContext
 from ..upscale import run_upscale_correction
 from ..utils import run_gc
 from .load import get_latents_from_seed, load_pipeline
@@ -21,7 +22,7 @@ logger = getLogger(__name__)
 
 
 def run_txt2img_pipeline(
-    job: JobContext,
+    job: WorkerContext,
     server: ServerContext,
     params: ImageParams,
     size: Size,
@@ -95,7 +96,7 @@ def run_txt2img_pipeline(
 
 
 def run_img2img_pipeline(
-    job: JobContext,
+    job: WorkerContext,
     server: ServerContext,
     params: ImageParams,
     outputs: List[str],
@@ -167,7 +168,7 @@ def run_img2img_pipeline(
 
 
 def run_inpaint_pipeline(
-    job: JobContext,
+    job: WorkerContext,
     server: ServerContext,
     params: ImageParams,
     size: Size,
@@ -217,7 +218,7 @@ def run_inpaint_pipeline(
 
 
 def run_upscale_pipeline(
-    job: JobContext,
+    job: WorkerContext,
     server: ServerContext,
     params: ImageParams,
     size: Size,
@@ -243,7 +244,7 @@ def run_upscale_pipeline(
 
 
 def run_blend_pipeline(
-    job: JobContext,
+    job: WorkerContext,
     server: ServerContext,
     params: ImageParams,
     size: Size,
