+++ api/onnx_web/worker/logging.py
@@ -1,0 +1,1 @@
+# TODO: queue-based logger
